requirements.txt
langchain
langchain-openai
langchain-chroma
unstructured[pdf]
python-dotenv
ragas
datasets
pandas

